Head Pack 3 ID changes
ID changes for Head Pack 3.

One minor note is that Human ID 142 is already listed as a PC, but I appear to have overwritten it with one of my heads.

Personally I think the new asset looks better anyway, and the impact is only one specific head out of dozens, so the likelihood of it affecting a live player is low.