Add vfx to space
Added multiple two new blaster bolt projectiles, one green and one red. They are both larger than the normal blaster bolt.

 For each combat module, a sound is played is played via vfx. This allows the sound to be consistently played without having to worry about the action queue.

Added a proton torpedo projectile for the missile launcher.

Added a vfx and sound for when a ship's shield or hull is hit (currently there is no vfx for when a hull is hit, I'll have to make that later on in a later update).