Implemented the BaseRecursiveLanguageTranslator class for all translators except droid and wookiespeak, as these don't follow a letter-based cipher yet. The other languages now translate according to their old definition + the general definitions.
Expanded on the general translations as to add all the planets known in star wars. Names including prime, III and such had only their first part added. Only when the planet name was long enough and not at risk of existing in a word.
Added the removal of successive defined characters from the output to the BaseRecursiveLanguageTranslator class; When according to the cipher definition z -> ' and ie -> 'a the output for zie would have been ''a, it will now instead be 'a. The basis for this is now only for the character ' but this could be overwritten or expanded for other languages, for example one using the character - in its output definitions often.