Hutlar Dungeon - Cleaned
This is a fresh PR because the previous one had the HoloNet and DM Broadcast PR folded into it by accident. This PR should only contain changes relating to the addition of the Hutlar Dungeon. This adds a new dungeon to Hutlar. Includes a limited set of Carapace Armor, which only has the armor, leggings, gloves and boots; it gives +30 poison defence each, and I am hesitant about having 8 pieces for +240 altogether. It also adds a few placeables, and includes the codes for a pitfall trap and the spawning of creatures upon a creature's death.