Enmity Rebalance Changes
Makes Tanking a little easier, and off-tanking (though require micromanaging) possible. This eases the gameplay off of Supports a little and makes fights, dependent on the presence of a tank, more chaotic as enemies will likely spread out without one. Will need a look at over the coming days to see if it's working well or not, and if not, a re-check.